Update day 18/04
I created a method to search and save albums in the database.

+++ app-user-api/src/main/java/br/com/sysmap/bootcamp/domain/entities/wallet/WalletEntity.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -26,7 +27,7 @@ public class WalletEntity {
     @Column(name = "user_id")
     private long userId;
 
-    private Integer balance;
+    private BigDecimal balance;
 
     @CreationTimestamp
     private LocalDateTime createdAt;
